capstone/requirements.txt
pandas
numpy
streamlit
json5==0.9.5
requests
datetime
matplotlib
seaborn
gspread
re
duckdb
gsheetsdb==1.0.4
google.oauth2
service_account